Only prefix with "SignavioXXX" if conflicting function name in core
