+++ kie-dmn/kie-dmn-feel/src/main/java/org/kie/dmn/feel/runtime/functions/BuiltInFunctions.java
@@ -95,36 +95,36 @@ public class BuiltInFunctions {
             new DayDiffFunction(),
             new HourDiffFunction(),
             new MinutesDiffFunction(),
-            new SecondsDiffFunction(),
+            new DiffFunction(),
             new YearAddFunction(),
             new MonthAddFunction(),
             new DayAddFunction(),
                                                                          new AppendAllFunction(),
                                                                          new ZipFunction(),
-                                                                         new SignavioNotContainsAnyFunction(),
-                                                                         new SignavioContainsOnlyFunction(),
-                                                                         new SignavioAreElementsOfFunction(),
+                                                                         new NotContainsAnyFunction(),
+                                                                         new ContainsOnlyFunction(),
+                                                                         new AreElementsOfFunction(),
                                                                          // TODO uncomment: new SignavioRemoveFunction(),
-                                                                         new SignavioRemoveAllFunction(),
-                                                                         new SignavioAvgFunction(),
-                                                                         new SignavioMedianFunction(),
-                                                                         new SignavioModeFunction(),
+                                                                         new RemoveAllFunction(),
+                                                                         new AvgFunction(),
+                                                                         new MedianFunction(),
+                                                                         new ModeFunction(),
                                                                          new SignavioEndsWithFunction(),
                                                                          new SignavioStartsWithFunction(),
-                                                                         new SignavioTextOccurrencesFunction(),
-                                                                         new SignavioTextFunction(),
-                                                                         new SignavioLeftFunction(),
-                                                                         new SignavioRightFunction(),
-                                                                         new SignavioMidFunction(),
+                                                                         new TextOccurrencesFunction(),
+                                                                         new TextFunction(),
+                                                                         new LeftFunction(),
+                                                                         new RightFunction(),
+                                                                         new MidFunction(),
                                                                          // TODO uncomment: new SignavioNumberFunction(),
-                                                                         new SignavioUpperFunction(),
-                                                                         new SignavioLowerFunction(),
-                                                                         new SignavioTrimFunction(),
-                                                                         new SignavioLenFunction(),
-                                                                         new SignavioIsSpacesFunction(),
-                                                                         new SignavioIsNumericFunction(),
-                                                                         new SignavioIsAlphanumericFunction(),
-                                                                         new SignavioIsAlphaFunction(),
+                                                                         new UpperFunction(),
+                                                                         new LowerFunction(),
+                                                                         new TrimFunction(),
+                                                                         new LenFunction(),
+                                                                         new IsSpacesFunction(),
+                                                                         new IsNumericFunction(),
+                                                                         new IsAlphanumericFunction(),
+                                                                         new IsAlphaFunction(),
             };
 
     public static FEELFunction[] getFunctions() {
