+++ kie-dmn/kie-dmn-feel/src/main/java/org/kie/dmn/feel/runtime/functions/RightFunction.java
@@ -16,24 +16,27 @@
 
 package org.kie.dmn.feel.runtime.functions;
 
+import java.math.BigDecimal;
+
 import org.kie.dmn.api.feel.runtime.events.FEELEvent.Severity;
 import org.kie.dmn.feel.runtime.events.InvalidParametersEvent;
 
-public class SignavioTrimFunction
+public class RightFunction
         extends BaseFEELFunction {
 
-    public SignavioTrimFunction() {
-        super("trim");
+    public RightFunction() {
+        super("right");
     }
 
-    public FEELFnResult<String> invoke(@ParameterName("text") String text) {
+    public FEELFnResult<String> invoke(@ParameterName("text") String text, @ParameterName("num_chars") BigDecimal num_chars) {
         if (text == null) {
             return FEELFnResult.ofError(new InvalidParametersEvent(Severity.ERROR, "text", "cannot be null"));
         }
+        if (num_chars == null) {
+            return FEELFnResult.ofError(new InvalidParametersEvent(Severity.ERROR, "num_chars", "cannot be null"));
+        }
 
-        String trimmed = text.trim();
-
-        String result = trimmed.replaceAll(" +", " ");
+        String result = text.substring(text.length() - num_chars.intValue(), text.length());
 
         return FEELFnResult.ofResult(result);
     }
